fix: Improve test options for all_models_register_on_metadata.
Set `connectable` attribute on alembic config inside this test. This
makes this test more compatible with documented `env.py` setup.

Add more input options for users to control how tests in all_models_register_on_metadata
are executed. Specifically, allow an `async_` flag to control which sort
of engine is produced into the `connectable` attribute, as well as an
`offline` flag to control whether the migration should run in offline
mode or not.