fix: Issue with runtime version_table_schema option.
`version_table_schema` appears to be (one of?) the only options that could
meaningfully affect the runtime behavior as-was, because it changes
where an un-configured (through env.py) MigrationContext instance would
point at an incorrect bit of data.

Now we run the process through env.py to ensure that setting it picked
up.

In order to attempt to avoid excess calls through the env.py, the
defaults tests now internally avoid making calls to `self.current`
wherever possible.